Renames spacebookingother to spacebooking
