+++ packages/app-store/apiHandlers.tsx
@@ -15,7 +15,7 @@ export const apiHandlers = {
   huddle01video: import("./huddle01video/api"),
   metamask: import("./metamask/api"),
   giphy: import("./giphy/api"),
-  spacebookingother: import("./spacebookingother/api"),
+  spacebookingother: import("./spacebooking/api"),
   // @todo Until we use DB slugs everywhere
   zapierother: import("./zapier/api"),
 };
