+++ packages/app-store/index.ts
@@ -12,7 +12,7 @@ import * as metamask from "./metamask";
 import * as office365calendar from "./office365calendar";
 import * as office365video from "./office365video";
 import * as slackmessaging from "./slackmessaging";
-import * as spacebookingother from "./spacebookingother";
+import * as spacebooking from "./spacebooking";
 import * as stripepayment from "./stripepayment";
 import * as tandemvideo from "./tandemvideo";
 import * as wipemycalother from "./wipemycalother";
@@ -33,7 +33,7 @@ const appStore = {
   office365video,
   slackmessaging,
   stripepayment,
-  spacebookingother,
+  spacebooking,
   tandemvideo,
   zoomvideo,
   wipemycalother,
