+++ packages/app-store/metadata.ts
@@ -11,7 +11,7 @@ import { metadata as metamask } from "./metamask/_metadata";
 import { metadata as office365calendar } from "./office365calendar/_metadata";
 import { metadata as office365video } from "./office365video/_metadata";
 import { metadata as slackmessaging } from "./slackmessaging/_metadata";
-import { metadata as spacebookingother } from "./spacebookingother/_metadata";
+import { metadata as spacebooking } from "./spacebooking/_metadata";
 import { metadata as stripepayment } from "./stripepayment/_metadata";
 import { metadata as tandemvideo } from "./tandemvideo/_metadata";
 import { metadata as wipemycalother } from "./wipemycalother/_metadata";
@@ -31,7 +31,7 @@ export const appStoreMetadata = {
   office365video,
   slackmessaging,
   stripepayment,
-  spacebookingother,
+  spacebooking,
   tandemvideo,
   zoomvideo,
   wipemycalother,
