+++ packages/app-store/spacebooking/_metadata.ts
@@ -8,8 +8,8 @@ export const metadata = {
   installed: true,
   category: "other",
   // If using static next public folder, can then be referenced from the base URL (/).
-  imageSrc: "/api/app-store/spacebookingother/spacebooking.webp",
-  logo: "/api/app-store/spacebookingother/spacebooking.webp",
+  imageSrc: "/api/app-store/spacebooking/spacebooking.webp",
+  logo: "/api/app-store/spacebooking/spacebooking.webp",
   publisher: "Cal.com",
   rating: 0,
   reviews: 0,
