+++ packages/app-store/spacebooking/package.json
@@ -4,7 +4,7 @@
   "name": "@calcom/spacebooking",
   "version": "0.1.0",
   "main": "./index.ts",
-  "description": "Looking to honor May 4th? Search no further. Download this app to make your booking success page resemble a long time ago in a galaxy far far away.",
+  "description": "Looking to honor May 4th? Search no further. Install this app to make your booking success page resemble a long time ago in a galaxy far far away.",
   "dependencies": {},
   "devDependencies": {
     "@calcom/types": "*"
