+++ packages/prisma/seed-app-store.ts
@@ -84,7 +84,7 @@ async function main() {
       api_key: process.env.GIPHY_API_KEY,
     });
   }
-  await createApp("space-booking", "spacebookingother", ["other"], "spacebooking-other");
+  await createApp("space-booking", "spacebooking", ["other"], "spacebooking-other");
   await createApp("zapier", "zapier", ["other"], "zapier_other");
   // Web3 apps
   await createApp("huddle01", "huddle01video", ["web3", "video"], "huddle01_video");
